Finish refactor.
