+++ solax/inverters/__init__.py
@@ -1,7 +1,7 @@
 from .qvolt_hyb_g3_3p import QVOLTHYBG33P
 from .x1 import X1
 from .x1_boost import X1Boost
-from .x1_boost_g4 import X1BoostG4
+from .x1_g4_series import X1G4Series
 from .x1_hybrid_gen4 import X1HybridGen4
 from .x1_mini import X1Mini
 from .x1_mini_v34 import X1MiniV34
@@ -25,7 +25,7 @@ __all__ = [
     "X3HybridG4",
     "X3",
     "X1Boost",
-    "X1BoostG4",
+    "X1G4Series",
     "X1HybridGen4",
     "X3MicProG2",
     "X3Ultra",
