+++ tests/fixtures.py
@@ -130,7 +130,7 @@ INVERTERS_UNDER_TEST = [
         method="POST",
         query_string="optType=ReadRealTimeData",
         response=X1_MINI_G4,
-        inverter=inverter.X1BoostG4,
+        inverter=inverter.X1G4Series,
         values=X1_MINI_G4_VALUES,
         headers=None,
         data=None,
@@ -170,7 +170,7 @@ INVERTERS_UNDER_TEST = [
         method="POST",
         query_string="optType=ReadRealTimeData",
         response=X1_BOOST_RESPONSE_G4_V3,
-        inverter=inverter.X1BoostG4,
+        inverter=inverter.X1G4Series,
         values=X1_BOOST_VALUES_G4_V3,
         headers=X_FORWARDED_HEADER,
         data=None,
